Minor bug fixes, added documentation, removed error pop up
