+++ frontend/src/components/popUpModal.svelte
@@ -3,7 +3,8 @@
     Contains pop-up modal
 
     Author: Brenda Dang
-    Last modified: 8/08/2024
+    Modified by: Bowen Dong
+    Last modified: 9/08/2024
 
 -->
 
@@ -17,9 +18,11 @@
     export let mainText = 'This is a pop-up modal.';  // Main text of popup modal
     export let firstButtonText = 'Button';  // Text for primary button (dark blue, or red)
     export let secondButtonText = 'Button';  // Text for secondary button (light blue)
-    export let iconPath = '';
-    export let width = '64';
+    export let iconPath = ''; // path for image icon
+    export let width = '64'; // default width of pop up
+
 </script>
+
 
 <div class="justify-center items-center font-sans h-full transition ease-in-out duration-300">
     <div class="fixed inset-0 flex items-center justify-center w-full h-full bg-black backdrop-blur md:bg-opacity-70">
@@ -55,7 +58,39 @@
                     />
                 </div>
             {/if}
-
         </div>
     </div>
 </div>
+
+<!--
+    EXAMPLE USAGE FOR ERROR POP UP
+<script>
+    import TopBar from "../../components/topbar.svelte"
+    import PopUpModal from "../../components/popUpModal.svelte";
+
+    let ifError = false
+
+    function dismissError() {
+        ifError = false;
+    }
+
+</script>
+
+
+<body>
+    <TopBar />
+    <button on:click={() => ifError = !ifError}>click to cause error</button>
+    
+    {#if ifError}
+        <PopUpModal 
+        header="Invalid audio format!"
+        mainText="Your meeting audio must be in MP3 or WAV format."
+        type='error'
+        iconPath='../src/assets/error-icon.svg'
+        firstButtonText="Re-upload"
+        firstHandleClick={dismissError}
+        width='96'/>
+    {/if}
+
+</body>
+-->
